# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/messaging/channels": {
      "get": {
        "operationId": "MessagingController_channels",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Canais de WhatsApp (tokens nunca expostos)",
        "tags": [
          "messaging"
        ]
      },
      "post": {
        "operationId": "MessagingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "provider": {
                    "type": "string",
                    "enum": [
                      "WHATSAPP_CLOUD",
                      "WHATSAPP_UNOFFICIAL_QR"
                    ],
                    "default": "WHATSAPP_CLOUD"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cria canal (oficial por padrão)",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/status": {
      "get": {
        "operationId": "MessagingController_status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Status e checklist de prontidão",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/onboarding/start": {
      "post": {
        "operationId": "MessagingController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inicia onboarding (Embedded Signup quando habilitado, senão conexão assistida)",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/onboarding/complete": {
      "post": {
        "operationId": "MessagingController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "wabaId": {
                    "type": "string",
                    "pattern": "^\\d{5,30}$"
                  },
                  "phoneNumberId": {
                    "type": "string",
                    "pattern": "^\\d{5,30}$"
                  },
                  "accessToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 1000
                  }
                },
                "required": [
                  "wabaId",
                  "phoneNumberId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Conclui conexão oficial (valida número↔WABA, inscreve webhooks)",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/connect": {
      "post": {
        "operationId": "MessagingController_connect",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reconecta/revalida credenciais armazenadas",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/disconnect": {
      "post": {
        "operationId": "MessagingController_disconnect",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desconecta (bloqueia envios imediatamente)",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/billing-status": {
      "get": {
        "operationId": "MessagingController_billing",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Status do faturamento próprio e histórico de verificações",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/billing-verification": {
      "post": {
        "operationId": "MessagingController_billingVerification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "method": {
                    "type": "string",
                    "enum": [
                      "API",
                      "ASSISTED"
                    ]
                  },
                  "evidenceNote": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "evidenceFileId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "confirmOwnBilling": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  },
                  "confirmNoCreditLineSharing": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  },
                  "approve": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "method",
                  "evidenceNote",
                  "confirmOwnBilling",
                  "confirmNoCreditLineSharing"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Envia evidência de faturamento próprio da WABA (revisão auditada)",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/templates": {
      "get": {
        "operationId": "MessagingController_templates",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Templates por evento",
        "tags": [
          "messaging"
        ]
      },
      "put": {
        "operationId": "MessagingController_upsertTemplates",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "templates": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "eventType": {
                          "type": "string",
                          "enum": [
                            "OS_RECEIVED",
                            "OS_DIAGNOSIS_STARTED",
                            "QUOTE_AVAILABLE",
                            "QUOTE_APPROVED",
                            "OS_WAITING_PARTS",
                            "OS_REPAIR_STARTED",
                            "OS_READY",
                            "OS_DELIVERED",
                            "WARRANTY_UPDATE"
                          ]
                        },
                        "name": {
                          "type": "string",
                          "pattern": "^[a-z0-9_]{1,512}$"
                        },
                        "language": {
                          "type": "string",
                          "pattern": "^[a-z]{2}(_[A-Z]{2})?$",
                          "default": "pt_BR"
                        },
                        "body": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "variables": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "maxItems": 10,
                          "default": []
                        },
                        "enabled": {
                          "type": "boolean",
                          "default": true
                        },
                        "approvalStatus": {
                          "type": "string",
                          "enum": [
                            "DRAFT",
                            "PENDING",
                            "APPROVED",
                            "REJECTED"
                          ],
                          "default": "DRAFT"
                        }
                      },
                      "required": [
                        "eventType",
                        "name",
                        "body"
                      ],
                      "additionalProperties": false
                    },
                    "maxItems": 50
                  }
                },
                "required": [
                  "templates"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cria/atualiza templates (variáveis permitidas; alteração exige nova aprovação Meta)",
        "tags": [
          "messaging"
        ]
      }
    },
    "/api/v1/messaging/channels/{id}/templates/sync": {
      "post": {
        "operationId": "MessagingController_sync",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sincroniza status de aprovação dos templates com a WABA",
        "tags": [
          "messaging"
        ]
      }
    },
...
    "/api/v1/messaging/test": {
      "post": {
        "operationId": "MessagingController_test",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "channelId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "to": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 30
                  },
                  "eventType": {
                    "type": "string",
                    "enum": [
                      "OS_RECEIVED",
                      "OS_DIAGNOSIS_STARTED",
                      "QUOTE_AVAILABLE",
                      "QUOTE_APPROVED",
                      "OS_WAITING_PARTS",
                      "OS_REPAIR_STARTED",
                      "OS_READY",
                      "OS_DELIVERED",
                      "WARRANTY_UPDATE"
                    ]
                  }
                },
                "required": [
                  "channelId",
                  "to",
                  "eventType"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Envia mensagem de teste real",
        "tags": [
          "messaging"
        ]
      }
    },
...
  },
  "info": {
    "title": "OrdemCerta API",
    "description": "API REST v1 — gestão de assistências técnicas. Erros: {code,message,details,requestId}.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}